IVYPORTAL-15837-Case-Widget-Complex-Custom-Fields
- split WidgetStateFilterBean from WidgetTextFilterBean
- split StateFilter from TextFilter
