+++ AxonIvyPortal/portal/src/com/axonivy/portal/bean/dashboard/WidgetTextFilterBean.java
@@ -16,15 +16,11 @@ public class WidgetTextFilterBean implements Serializable {
   private static final long serialVersionUID = 3218284115015773931L;
 
   private static List<FilterOperator> operators = FilterOperator.TEXT_OPERATORS.stream().toList();
-  private static List<FilterOperator> stateOperators = FilterOperator.STATE_OPERATORS.stream().toList();
 
   public List<FilterOperator> getOperators() {
     return operators;
   }
 
-  public List<FilterOperator> getStateOperators() {
-    return stateOperators;
-  }
 
   public void onChangeOperator(DashboardFilter filter) {
     
@@ -43,16 +39,4 @@ public class WidgetTextFilterBean implements Serializable {
    return false;
   }
 
-  public boolean isShowStateListPanel(DashboardFilter filter) {
-    for (FilterOperator operator : stateOperators) {
-      if (operator == FilterOperator.EMPTY || operator == FilterOperator.NOT_EMPTY) {
-        continue;
-      }
-      if (operator == filter.getOperator()) {
-        return true;
-      }
-    }
-
-   return false;
-  }
 }
