+++ AxonIvyPortal/portal/src_hd/com/axonivy/portal/dashboard/component/filter/DashboardFilter/DashboardFilter.xhtml
@@ -43,7 +43,7 @@
     </h:panelGroup>
 
     <h:panelGroup layout="block" rendered="#{filter.isState()}">
-      <ic:com.axonivy.portal.dashboard.component.filter.TextFilter filter="#{filter}"
+      <ic:com.axonivy.portal.dashboard.component.filter.StateFilter filter="#{filter}"
         componentToUpdate="#{cc.attrs.componentToUpdate}" filterIndex="#{cc.attrs.filterIndex}"
         readOnly="#{readOnly}" messageId="#{cc.attrs.messageId}" />
     </h:panelGroup>
