+++ AxonIvyPortal/portal/src_hd/com/axonivy/portal/dashboard/component/filter/StateFilter/StateFilter.rddescriptor
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
+<richDialogDescriptor>
+    <property>
+        <name>viewTechnology</name>
+        <value>JSF</value>
+    </property>
+</richDialogDescriptor>
