+++ AxonIvyPortal/portal/src_hd/com/axonivy/portal/dashboard/component/filter/StateFilter/StateFilterData.ivyClass
@@ -1,0 +1,2 @@
+StateFilterData #class
+com.axonivy.portal.dashboard.component.filter.StateFilter #namespace
